+++ src/main/java/com/online/shelter/pet/servlet/model/AnimalWithDownplayWeight.java
@@ -10,7 +10,7 @@ public class AnimalWithDownplayWeight {
     private final char sex;
     private final String color;
     private final double age;
-    private final double growth;
+    private final int growth;
     private final double weight;
 
     private final String namePerson;
@@ -19,7 +19,7 @@ public class AnimalWithDownplayWeight {
 
     private final boolean downplayWeight;
 
-    public AnimalWithDownplayWeight(LocalDateTime dateTime, String typeAnimal, String nameAnimal, String breed, char sex, String color, double age, double growth, double weight, String namePerson, String phone, String email, boolean downplayWeight) {
+    public AnimalWithDownplayWeight(LocalDateTime dateTime, String typeAnimal, String nameAnimal, String breed, char sex, String color, double age, int growth, double weight, String namePerson, String phone, String email, boolean downplayWeight) {
         this.dateTime = dateTime;
         this.typeAnimal = typeAnimal;
         this.nameAnimal = nameAnimal;
